#1264: net.c moved back to studio
